feat(agents): port frontmatter validation to CLI (STORY-025, GH #193)
Port P-003 Agent tool restriction check from standalone script to CLI
handler. Resolve H-10 violation by splitting 3 classes into own files.

- Delete scripts/validate-agent-frontmatter.py (403 lines)
- Delete scripts/tests/ directory entirely
- Split ValidateFrontmatterCommand into 4 files (H-10 compliance)
- Port P-003 check into ValidateFrontmatterCommandHandler
- Add tests/unit/agents/test_p003_agent_tool_restriction.py (8 tests)
- Update CI to single validation step
