--- scripts/tests/__init__.py
@@ -1,10 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# Copyright (c) 2026 Adam Nowak
-
-"""
-Guardrail Hook Tests
-
-Work Item: WI-SAO-015
-
-This package contains BDD tests for the guardrail hook system.
-"""
